Split Hunt Scan off from Listener
+ check for duplicate method
